Change values to float

+++ app/Helpers/InsightsHelper.php
@@ -4,12 +4,12 @@ namespace App\Helpers;
 
 class InsightsHelper
 {
-    public static function incrementGauge(string $name, array $labels = [], int $increment = 1): void
+    public static function incrementGauge(string $name, array $labels = [], float $increment = 1): void
     {
         app('prometheus')->getGauge($name)->incBy($increment, $labels);
     }
 
-    public static function decrementGauge(string $name, array $labels = [], int $decrement = 1): void
+    public static function decrementGauge(string $name, array $labels = [], float $decrement = 1): void
     {
         app('prometheus')->getGauge($name)->decBy($decrement, $labels);
     }
